puts editor in its own component
