+++ src/components/Formula/FormulaPage.js
@@ -7,61 +7,15 @@ import {
   InputLabel,
   Typography,
 } from "@material-ui/core";
-import React, { useEffect, useState } from "react";
-import AceEditor from "react-ace";
-import "ace-builds/src-noconflict/mode-ruby";
-import "ace-builds/src-noconflict/theme-monokai";
+import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
 
 import "./editor.css";
-import languageTools, {
-  setCompleters,
-  addCompleter,
-} from "ace-builds/src-noconflict/ext-language_tools";
 import FormulaTester from "./FormulaTester";
 import { update } from "lodash";
-
-const Editor = ({ value, availableVariables, updateExpression }) => {
-  const sendExpressionToParent = newValue => {
-    updateExpression(newValue);
-  };
-
-  useEffect(() => {
-    setCompleters([languageTools.snippetCompleter]);
-    addCompleter({
-      getCompletions(editor, session, pos, prefix, callback) {
-        const results = availableVariables.map(availableVariable => {
-          return {
-            caption: availableVariable,
-            value: availableVariable,
-            meta: "available variable",
-          };
-        });
-        callback(null, results);
-      },
-    });
-  }, [value, availableVariables]);
-  return (
-    <AceEditor
-      mode="ruby"
-      theme="monokai"
-      fontSize={16}
-      onChange={event => sendExpressionToParent(event)}
-      value={value}
-      wrapEnabled
-      minLines={10}
-      width="100%"
-      height="200px"
-      setOptions={{
-        enableBasicAutocompletion: true,
-        enableLiveAutocompletion: true,
-        enableSnippets: false,
-      }}
-    />
-  );
-};
+import FormulaEditor from "./FormulaEditor";
 
 const Formulas = ({ label, formulas }) => {
   return (
@@ -89,6 +43,7 @@ const Formulas = ({ label, formulas }) => {
     </>
   );
 };
+
 const useStyles = makeStyles(theme => ({
   formControl: {
     minWidth: 240,
@@ -175,7 +130,7 @@ const FormulaPage = ({
           </Grid>
 
           <Grid item>
-            <Editor
+            <FormulaEditor
               value={formulaToUse.expression || ""}
               availableVariables={availableVariables}
               updateExpression={updateExpression}
